Move samples.py to plugin/, move hardware scripts to plugin/scripts/, delete launch_napari.py and `repeat-cal-acq-rec.py`
