+++ tests/widget_tests/test_sample_contributions.py
@@ -1,4 +1,4 @@
-from waveorder.scripts.samples import download_and_unzip
+from waveorder.plugin.samples import download_and_unzip
 
 
 def test_download_and_unzip():
