+++ waveorder/napari.yaml
@@ -10,13 +10,13 @@ contributions:
     python_name: waveorder.io._reader:napari_get_reader
   - id: waveorder.polarization_target_data
     title: Polarization Target Data
-    python_name: waveorder.scripts.samples:read_polarization_target_data
+    python_name: waveorder.plugin.samples:read_polarization_target_data
   - id: waveorder.polarization_target_reconstruction
     title: Polarization Target Data
-    python_name: waveorder.scripts.samples:read_polarization_target_reconstruction
+    python_name: waveorder.plugin.samples:read_polarization_target_reconstruction
   - id: waveorder.zebrafish_embryo_reconstruction
     title: Zebrafish Embryo Reconstruction
-    python_name: waveorder.scripts.samples:read_zebrafish_embryo_reconstruction
+    python_name: waveorder.plugin.samples:read_zebrafish_embryo_reconstruction
   readers:
   - command: waveorder.get_reader
     accepts_directories: true
